[split] don't change the value of 'confirmOpenNewProject' option in remote development sessions (RDCT-1505, GTW-9099)
ProjectLifecycleUiCustomization service is introduced to allow always opening projects in a new window without modifying the option. This way we ensure that users won't be able to change the value of this option manually after it set in code in remote dev mode. Also, running the IDE in remote dev mode won't affect settings for monolith (full) IDE.

GitOrigin-RevId: 794486d9146a616277933585ab9cc32485414bcc