[split] don't change the value of 'reopenLastProject' option in remote development sessions (RDCT-1554)
ProjectLifecycleUiCustomization.canReopenProjectOnStartup property is introduced to allow disabling reopening projects on startup without modifying the option. This way we ensure that users won't be able to change the value of this option manually after it set in code in remote dev mode. Also, running the IDE in remote dev mode won't affect settings for monolith (full) IDE.

GitOrigin-RevId: 88900647ab1e9acebde48f7d8afbd37816e94ebd