IJPL-157074 Support UnixPtyProcess in UnixProcessManager.getProcessId
This reflection mess relies on the exact class name. Therefore,
it doesn't work for custom Process implementations. While we
still support Java 8 here, the only really safe thing we can do
here to support UnixPtyProcess is to explicitly add it to the pid()
white list. It'll do for now until we finally get rid of Java 8
legacy, which should be soon enough already.

GitOrigin-RevId: 226db5fe288ecd8a1f80777fc5c6ade9e5f51aeb