IJPL-43794 Add Search Everywhere step to New Users Onboarding
Also, have to fix the regex used to parse the shortcut definition in the GotIt tooltip. It was skipping the `Shift` keystroke because it is converted to `pressed SHIFT` and the space inside it was not recognized.

(cherry picked from commit e1a9e37bf204454be281e62dc33c94094a9e78cd)

IJ-CR-140970

GitOrigin-RevId: 722a206eff97890a5b923e361cc7243ce38f7252