[split] Make selection listener works in split-mode.
[split] Make selection listener works in split-mode.

More context: in split mode we decompose selection events for a tree and the host got the following events:
1. Add a line to selection
2. Remove the previous line from selection.

[split] Make selection listener works in split-mode.

More context: in split mode, we decompose selection events for a tree, and the host got the following events:
1. Add a line to selection
2. Remove the previous line from selection.
For the second event, we get `e.isAddedPath() == false` and we don't update console view. After my changes console view will be updated properly during the second event.


Merge-request: IJ-MR-139041
Merged-by: Sviatoslav Vlasov <Sviatoslav.Vlasov@jetbrains.com>

Merge-request: IJ-MR-139377
Merged-by: Sviatoslav Vlasov <Sviatoslav.Vlasov@jetbrains.com>

GitOrigin-RevId: 7ded4281da1cbc468ffde3d8a916743aeb54b955