[kotlin] don't yield KaModule based on sdkEntry if sdk is not attached to the context module
In multi-SDK project, jdk classes are found in all those SDKs. Now, if all sdks produce KaModules, then in some IJ modules, the first KaModule in enumeration (`ModuleInfoProviderUtils.firstOrNull(ModuleInfoProvider, PsiElement, ModuleInfoProvider.Configuration)``) would be wrong (corresponds not to the sdk, attached to that module). And this makes `LLFirSelectingCombinedSymbolProvider.selectFirstElementInClasspathOrder` to skip that sdk module during jdk's java class search because the module is not in the dependent map.

Result: red code on java classes from jdk

^KTIJ-30560 fixed

GitOrigin-RevId: 7dfd96a11c40f3be37a5e9ab73efd5df0aad7274