IJPL-157074 Prohibit sending signals to PID<=0
Doing so creates very nasty bugs that are incredibly
hard to debug, because sending something like SIGINT
to PID=0 kills every single user process almost with
no traces except perhaps in the OS log.

To avoid repeating such bugs in the future, simply
prohibit this. It's not like we actually ever need
this functionality in the platform.

GitOrigin-RevId: 2f15d28ce9f6fa97793e030696c80f91057a4d54