IJPL-157074 Temporarily fix pty4j stdout/stderr blocking issues
The new UnixPtyProcess implementation for macOS on Intel
in pty4j doesn't correctly signal I/O streams blocked
on reading from the process, so they end up blocking
forever after the process has finished.

There's already a method that does the job, but it isn't
called in pty4j, so to fix this issue for now we call
it ourselves using reflection.

GitOrigin-RevId: d7863c9151274a769b0197f27aa68a3dab11b431