IJPL-43794 Refactor: create a separate module for New Users Onboarding
And bundle it only to IDEA for now.
Separate module is needed because later we need to add the dependency on `intellij.platform.experiment` to implement the A/B experiment.
We can't add this dependency to the existing `intellij.platform.ide.newUiOnboarding` module because `intellij.platform.experiment` is bundled only in IDEA now. So then New UI Onboarding will be present only in IDEA, while it also should be in the other IDEs.

(cherry picked from commit 3dc524b97d7ef6667e0062db5d76aa1f3b56d12f)

IJ-CR-140970

GitOrigin-RevId: 6d8cb5a80e45fc79b6418707c4aea8aaa9d7260f