IDEA-355257 [TestFramework] fix: removed usage of the Junit 4 comparison failure inside the ComparisonFailureData
The ComparisonFailureData cannot depend on the Junit 3-5 classes. It should be universal over all Java test frameworks. Also, artifacts for these Junit classes existed only in compile time. That is the reason of hidden class is not found exception.

The deprecated FileComparisonFailure extends the FileComparisonData for two years. This interface is covered by the ComparisonFailureData.createFileComparisonData function. Also, the old one ComparisonFailureData.createFileComparisonFailure converts the FileComparisonFailure using the functions from the FileComparisonData interface. Therefore, we can remove the createFileComparisonFailure function without any functional loose.

Ps. The FileComparisonFailure planned to be removed in IDEA 2025.1 (IJPL-156277).

GitOrigin-RevId: 2aa1967fe0732d2add105020aab56fbc65cb3d6c