[core] IJPL-3459: prohibit adding new ShutDownTracker tasks if shutdown hook is already started
In many cases we rely on shutdown tasks to be executed in specific order -- in order we've added them to ShutDownTracker: e.g. shutdown task 1 added before shutdown task 2 => we expect shutdown task 2 to be executed first (LIFO), and shutdown task 1 after it.

But if shutdown has initiated very early in an app lifecycle, that assumed order could be violated because task 1 already added to the tracker, and already executed, while task 2 is not yet added, and it is added later, and gets executed later.

To prevent these races, I prohibit adding more tasks to the tracker if shutdown is already started -- `ProcessCanceledException` is thrown in this case

GitOrigin-RevId: 07611dc3ac94c7697bb7dd4b8aece5fb76e9e1ce