IJPL-158384 Fix Expand All tooltip
A simple bug, but a double one. First, the getShortcutText
method accepting an action ID doesn't return null or an empty
string if there's no shortcut. It returns "<no shortcut>" instead.
Second, there's no escaping HTML characters here which is
why even that string is not displayed, as it's interpreted as
an invalid tag.

Fix both. Use the overload accepting a keyboard shortcut
and check that the shortcut exists before that. This handles
the "no shortcut" case. Escape the shortcut text before passing
it into the tooltip, just in case we have shortcuts like "Ctrl + <"
or something like that.

GitOrigin-RevId: a73af6ef6535812aa83f3b098ab09131679fa968