Add segmented controls (#387)
* Add segmented controls

* Fix colors and focus behaviour of segmented controls

* Modify focus behaviour and colors of segmented controls

* Add segmented controls

* Fix colors and focus behaviour of segmented controls

* Modify focus behaviour and colors of segmented controls

* Tweak implementation to match actual impl/behaviour

The one thing that is still TODO is avoid focusing the control when it's
clicked. I removed the previous "clear focus" behaviour because it
clears the focus on click even when it shouldn't. I'll come up with a
better solution as soon as I can.

* Change focus behaviour to match Swing's

We made the buttons not focusable anymore, to prevent unwanted focus
movement. There is a bit of focus dark magic, but it now works as we'd
expect.

* Update segmented controls api

---------

Co-authored-by: Sebastiano Poggi <sebp@google.com>
GitOrigin-RevId: 3e847ba84adc08a489518d73ecdc8bd9410b483b