[Workspace Model] [IDEA-348672] Use concurrent hash map to store the information about the unloaded modules
Generally, it seems like this can be rewritten to just use the unloaded storage and do not store the additional information in maps.
This change is intended to fix the concurrency issue.

With the change of the map, we loose the ordering of the keys. The ordering, however, doesn't matter. It's only exposed via `allModuleDescriptions` and `unloadedModuleDescriptions` functions, but they have a `Collection` type that doesn't specify the ordering. Also, in practice, the ordering is not used not by intellij, nor by external plugins code.

The map of modules was original unordered and ordering was added in 5f7c407aae45df034a3dc1c0a9876a53807b64b8 without any specific reason for that.

GitOrigin-RevId: 8e3f0a25fc14e1315127cca88d7658bd7d6e5d59