[Workspace Model] [IJPL-14746] Remove printing stack trace in when the cache is not loaded
The cache not loading because of some file-related issues is a normal situation. We print to log a warning with a stack trace saying that deserialization is not available.

However, having a stack trace doesn't give a lot of information because we don't load cache from many different places, and also it bothers some users because they this this is an actual exception.

Now we print only the message from the exception and print the stacktrace in case of debug logging

GitOrigin-RevId: 9349a32c8f4f8c50b2ab2617911bc44d8351e24d