IJPL-158590 Ignore subsequent attempts to set a cached tree presentation
Once a presentation is set, it's updated as tree nodes are loaded.
If we attempt to set it again, the new presentation won't have this
cached data.

We could try to recreate or copy that data, but it's not necessary
because if the presentation is set so quickly, it's likely the same one.
So we can just ignore it.

GitOrigin-RevId: 8fe4fd55658e141baab0a581a983509b35989c37