IJPL-157156 Make the editor tabs at least as high as tool window headers
Not clear what caused this regression, but there's an easy way to both
fix it and prevent it. We just need to make sure that the editor tabs
are always at least as high as tool window headers.

The only trick is that the height is calculated in JBTabsImpl, which
is used not just for the editor tabs. So we add an internal protected
method here and override it in the editor tabs only.

GitOrigin-RevId: 25820e7c639d84e0d76f40acd2b8c7038e46773c