[jps build] use relative paths when loading the workspace model from cache if and only if IDE uses relative paths (IJPL-157801)
StorageSerializerUtil.getVirtualFileUrlSerializer stores VirtualFileUrl instance in completely different format if 'ide.workspace.model.store.relative.paths.in.cache' registry option is enabled. Deserialization in the build process must use the same format as the IDE, so now value of this registry options is passed as 'ide.workspace.model.store.relative.paths.in.cache' system property to the JPS build process and used there to switch relativization on or off.

GitOrigin-RevId: a87b7f4f1a2c073a8d1fdaaa72a38b5dc11d0e01