[jps build] use relative paths when loading the workspace model from cache if and only if IDE uses relative paths (IJPL-157801)
StorageSerializerUtil.getVirtualFileUrlSerializer stores VirtualFileUrl instance in completely different format if 'ide.workspace.model.store.relative.paths.in.cache' registry option is enabled. Deserialization in the build process must use the same format as the IDE, so now value of this registry options is passed as 'ide.workspace.model.store.relative.paths.in.cache' system property to the JPS build process and used there to switch relativization on or off.

GitOrigin-RevId: f5bd3adf45cf54fdf8dc9eedf73b55d8f6668a00