IJPL-157074 Add the run.processes.using.pty.helper.on.mac.intel key
This way we can easily disable this hack if something goes wrong.
Do note that while we can disable this flag, it naturally won't
revert the pty4j version to the pre-hack state, so it's not a 100%
replacement for reverting all the commits that introduced
this hack to begin with.

GitOrigin-RevId: c6b0308cf031ecde71702f8eaf1d7512e22db0f9