[external system][IDEA-355927] do not migrate project classpath between different versions of the persistent state
The `CACHE_FILE` for `PersistentStateComponent` never migrates between IDE versions. However, a dirty cache with a non-empty `projectBuildClasspath` will cause a deadlock on `GradleSettings` initialization if the initial call on settings is made from EDT. This is because `loadState` does not pump EDT; as a result - any lock executed from `loadState` will cause a deadlock.

GitOrigin-RevId: eba5263068304c670c364eedd862cdaa9d85b8c1