[RDCT] RDCT-1631, IJ-CR-141431: Implement `noStateLoaded` for `ConsoleFoldingSettings` to fix synchronization of the default state
The common problem with settings synchronization may be settings components that don't implement `noStateLoaded` method.
When settings are loaded in the default state, the actual value of the component state is usually `null`. In such cases, `noStateLoaded` is called instead of `loadState`. If this method is not implemented, then such values will not be synchronized.

(cherry picked from commit 7126989f53932dfdcf282901476952c3a96a4d09)

GitOrigin-RevId: 2503127df7d188db5797707758d1e053a5430174