KTIJ-25346 [kotlin] Handle typealiased constructors in K2 Import Optimizer
Currently, due to KTIJ-26098, if the constructor is invoked via the typealias,
the reference resolves directly to the constructed class, and not to the
typealias.

This prevented `UsedReferencesCollector` from properly recognizing such references.

To handle this, we resolve the type explicitly when we see constructor calls
by using code fragments.
This works because type references work fine with typealiases (compared
to constructor references).

Also, add an explicit test for currently incorrect optimization
of `kotlin.concurrent.Volatile` import.

^KTIJ-25346 Fixed

GitOrigin-RevId: f531c00cba6d7e6687a45ae06aa9ac85a92262a3