PY-74236: Move SDK saving logic to the service, create presenter for the "create SDK" dialog, accept callback on SDK creation.
`PyActiveSdkConfigurable` has `model` that must be in sync with the JDK table.

After SDK addition, `PyActiveSdkConfigurable.updateSdkListAndSelect` must be called to select the active SDK.

The user then clicks "apply," see `PyActiveSdkConfigurable.apply` to associate with a model.

`PythonAddLocalInterpreterDialog` delegates SDK creation to `PythonAddEnvironment` (so-called `currentManager`).

It then calls `PythonAddLocalInterpreterDialogPresenter` that acts as presenter to persistent SDK using `PySdkService` and reports SDK using flow which is connected to `PyActiveSdkConfigurable.updateSdkListAndSelect`.

GitOrigin-RevId: 4114fe3385f8b51f82d6465617c27fbe65e8ab04