[performanceTests] Don't close SE on focus lost in perf tests
Otherwise, the command doesn't function properly since the following happens:
1. We open SE, type something, select first file
2. We open SE again but in the meantime file is opened and takes the focus, we close SE

This case can be fixed by subscription to FileEditorManagerListener but the case when we select already opened file can't. Since in this case, there is no listener or callback that will indicate that the focus is moved to the editor.

GitOrigin-RevId: 0e7959443fcd823ff9ddf10b121fa38e257f38cf