[python] fix a bug with obtaining ancestor classes in `com.jetbrains.python.psi.impl.PyClassImpl.mroLinearize`
Fix a bug with a situation when two identical inctances of type `Generic` were stored in the map in `com.jetbrains.python.psi.impl.PyClassImpl.mroLinearize` and were processed separately due to not matching hashcodes and equals methods. This situation caused an `MROException` to be thrown so the class ancestors could not be defined properly. It affected generic types matching as an actually generic class was not recognized as generic because of the wrong list of ancestors.

GitOrigin-RevId: 391390c5143b0c7247bd55f148d519f80063d5a8