remove obsolete highlightings in case of overlapping sessions
when highlighting is canceled too frequently, two HighlightingSessions could overlap for short amount of time and fight for recycling the same highlighter. The old session recycler grabbed the highlighter, the new session tried to recycle it again but failed, the old one canceled and restored the highlighter to the recyclable state, the new session proceeded with creating new highlighter. Thus, two highlighters are created and one of them was never recycled.

GitOrigin-RevId: 1bdd6acb72b0182b1af1ea507e8809536434ca26