`intellij.platform.kernel` as a simple module (rather than V2 plugin)
This module is a core component of Rhizome - it's required for any interaction with Rhizome DB, creating entities, etc.
It was a V2 plugin as we wanted to make sure that only new V2 plugins will be able to use this.

We expected that it won't be much of a struggle, considering that every feature that would want to use Rhizome is going to be split into shared/backend/frontend modules (which could be V2 from the start).
However, we faced the issue that it's not possible to move some public API to V2 modules because it breaks backward compatibility, and therefore this API cannot depend on V2 module.

The first such issue was rewriting the progresses on Rhizome (RDCT-1444).
It's possible to create some interlayers to make it work without breaking compatibility and keeping kernel V2, but it was decided that it isn't worth it in such cases.

GitOrigin-RevId: 0488ca905bd145b2c3b6b2fa8481fa59b9f76529