[kotlin] add a list of plugins which are safe to enable in K2 IDE mode
Currently, if a plugin depends on Kotlin plugin in K2 mode and the plugin is not marked as k2-compatible, then it's going to be disabled. We have around 550 plugins depending on Kotlin. A lot of them do not use resolve. This commit adds a list of plugins which are safe to enable in K2 mode.

KTIJ-30545
^KTIJ-30546 fixed


(cherry picked from commit 3707395411167ef2ca12693f97a40a0f75b95357)

IJ-MR-141637

GitOrigin-RevId: bb1f21debb4fba3ae735e640c89fb54e6904637d