[rider + json] IJPL-159354 Add explicit JSON plugin dependency to all plugins whose modules contain compile dependency on the JSON module
The assumption here is: if there is an explicit compile dependency on the JSON module, the corresponding plugin indeed relies on it -> it requires the JSON _plugin_ to be installed and enabled.

This might be an overkill, but currently JSON is bundled into all IDEs -> should not make any existing plugin unavailable because of unsatisfied dependency. Should anybody decide that a particular plugin does not need a JSON dependency, they can remove it from the corresponding plugin.xml and JPS module (and fix possible compilation errors).

GitOrigin-RevId: b10774cc51a4e195b2ee4dc7a5ad1bcef40444a9