IJPL-159009, IJPL-159805 Fix collect() modality in showingScope()
The coroutine that starts collecting values must react to them
regardless of any modalities. This is because showingScope()
can be invoked when the component is created, but later the component
may be added to some modal dialog, which can be shown even later.
In this case, the collector must react and start the component
coroutine with the proper modality.

The current modality when the "outer" coroutine is launched
doesn't matter at all because a component can be created
at any moment. It's the modality of the component when
it's actually shown that matters. But it's already handled by
the "inner" coroutine.

GitOrigin-RevId: e8bc0e1eecfcba2d91863f646fa9ef2f23c997ac