IJPL-43686 Fix stuck resize cursor
Because on macOS the cursor may be updated asynchronously,
sometimes it gets stuck in a state when WindowResizeListenerEx
"thinks" it's already reset it to the default one, while in reality
it's some kind of resize cursor. This looks like the resize cursor
is everywhere, and it's stuck like that until the popup is closed.

Fix by moving the setting of the cursor on the component
outside the if check. There's another check inside that
compares the cursor to the actual component's cursor,
so there's no performance penalty here.

GitOrigin-RevId: f22613339249e833d72058ae383b781a313be836