IJPL-43686 Implement a workaround for macOS mouse entered/exit bug
Sometimes mouse entered/exit events never arrive to the popup
when using rounded corners. The events in question never even
reach the AppKit thread, so it looks like an AppKit bug.
This results in sun.lwawt.LWWindowPeer.lastCommonMouseEventPeer
being null, because the mouse "leaves" whatever is outside the popup
(usually the IDE frame), but never "enters" the popup itself.
This in turn breaks sun.lwawt.LWCursorManager.updateCursorImpl
because it relies on the current peer to find the component
to get the cursor from.

Work around this bug by sending a fake MOUSE_ENTERED event
when a mouse move event is received, but the peer doesn't
match the popup's peer. Some reflection hacks are used to
access macOS-specific internals.

GitOrigin-RevId: df7e05f72b53049fc50d77a4d34d400c53448293