KTIJ-30908 [kotlin] Do not propose `ChangeVisibilityFix` on `KtParameter` unless it has `val/var`
Also, skip all the private declarations on the way to the parent declaration

N.B. `exposedConstructorParameterType_val.kt` doesn't work in K1, because there
is a more advanced check which checks if the changed visibility modifier will
actually make the problem go away
(see `ExposedVisibilityChecker.checkDeclarationWithVisibility`).

In the case of the constructor value parameter, the exposure error will stay,
but it can be fixed by also making the constructor private with a second invocation
of the quickfix.

^KTIJ-30908 Fixed

GitOrigin-RevId: ff4207a3dc71b7c185112aa128a13e10c38fd895