IJPL-159317 IDE highlights syntax in the comment
rework the HighlightInfoUpdater to avoid obsolete error elements stuck in the markup
The problem was in the several highlighting sessions overlapping each other when the one session was canceled but still lingered on while the other started and progressed far enough.
The two sessions called setHighlightersInRange() in HighlightInfoUpdater in reentrant fashion which could lead to some highlighters to be forgotten by HighlightInfoUpdater internal data structures. The fix is to refactor HighlightInfoUpdater to guarantee that the highlighters registered there are always contained in the markup and vice versa.
Several testcases added to DaemonRespondToChangesTest where the test invokes typing, then waits for the daemon to start, then proceed to another typing, effectively canceling the daemon process very frequently to induce the session overlappings and then checking that no error element highlighters are left in the markup

GitOrigin-RevId: 9b72068d71b28d940c3e9fdb98574c78e4776800