Disable task-based UI in game tools
When task-based UI is enabled, the device and process views are
hidden, making it impossible for a user to select a process.

Even if the user could select a process, the main profiler view
would not appear, preventing the user from profiling the process.

This disabled the task-based UI via a command line flag.

Bug: 338285051
Change-Id: I035c343208472044a1910499a109d7af245b2d24