IJPL-159968 Save and log systemScaleFactor first access stack trace
When we get a "Must be not computed before that call" error,
it's very hard to reason about the cause. It can be computed
almost anywhere, in any code that calls JBUIScale.systemScaleFactor,
often indirectly, by accident.

Save the stack trace the first time systemScaleFactor and later
attach it as the cause to the exception that's logged when
one of the preload() functions is called.

GitOrigin-RevId: 02eadede9353d2c6b862590881a9d8aafe11e541