[gateway] use the proper client build number when launching the client for a backend with a snapshot version (GTW-6851)
When a backend has a snapshot version, CodeWithMeClientDownloader.createSessionInfo computes the latest available build of the client, and it is downloaded. However, code which launches the client used 'xxx.SNAPSHOT' version and failed to run it. Now the exact version of the client is propagated via JetBrainsClientDownloadInfo.clientBuildNumber and ExtractedJetBrainsClientData.version and used to run the downloaded client.

GitOrigin-RevId: bfceb6fc4e40ef4b7177ba109615a1250b355eb7