IJPL-159968 Add JreHiDpiUtil.preload to JBUIScale.preload
Similarly to how JBUIScale.systemScaleFactor can be computed
incorrectly and then cached, the same thing can happen to
JreHiDpiUtil.isJreHiDPIEnabled, which in turn will cause
the computation of systemScaleFactor to go wrong
(for example, isJreHiDPIEnabled relies on
SunGraphicsEnvironment.uiScaleEnabled_overridden
which is initialized along with LaF).

Add a similar assert to JBUIScale.preload to check that the value of
JreHiDpiUtil.isJreHiDPIEnabled was not initialized before. Which it
normally shouldn't. Except when it's called from initLux(),
which invokes static initialization of UiUtil, which calls JBUI.scale,
and that causes early init of isJreHiDPIEnabled. For Lux it seems
to be OK, though, just that we have to take that into account,
which is why the new JreHiDpiUtil.preload is allowed to
be called several times, ignoring the subsequent calls,
as to not throw an error if it was preloaded intentionally,
as opposed to an accidental JreHiDpiUtil.isJreHiDPIEnabled
call somewhere.

JreHiDpiUtil.isJreHiDPIEnabled had to be refactored a bit to ensure
that all computations happen inside initialization. No more bypass
for macOS and Wayland. Technically, on macOS we could simply
skip the check in JBUIScale.preloadOnMac, but it's still better
to have it, as detecting early access on macOS, where it's harmless,
could help us to fix it on other platforms, where it isn't.

GitOrigin-RevId: 31989284e85cfd09903f79debdf5858e0c0d0177