PY-75188: Obey allowed env types when restoring preferred envs, so Jupyter uses Conda only.
Each project type might have a list of allowed envs. For Jupyter this is conda only.
We store user choice and restore it back on the next new project trying to suggest the same env type user used previously.

However, we must limit this by a list of allowed interpreters: even if vanilla CPython was used for the previous project, we still can't use it for Jupyter.

GitOrigin-RevId: cb0e5b311d1f2020df474df9fe436875a386f55c