[rdct] import settings from the previous version when the IDE backend starts for the first time (GTW-7436)
Implicit switching off config importing in remote dev mode is removed (it was added in 2e962a2b8851c6, and the reasons for that are unclear).
ConfigImportHelper is changed to ensure that Application isn't accessed during importing in remote dev mode.
Since Gateway first starts the backend with a headless command (cwmHostStatus), and importing doesn't happen while executing headless commands (see IJPL-160167), `merge-configs` is saved to `migrate.config` file to ensure that migration will be started when the backend runs after executing a headless command. Currently it's done only for commands related to remote dev to avoid changing the behavior of regular IDEs, later we can do this for all commands.

GitOrigin-RevId: cf701b817ed041f11c1a2ea3059b83f1859a6029