IJent: don't use `CorePosixFilteringFileSystemProvider` in `MultiRoutingFileSystemProvider`
Sometimes this wrapper brings problems. For example, `org.apache.commons.compress.archivers.tar.TarArchiveEntry.readOsSpecificProperties` checks for "posix" attribute and calls a branch that tries to get `PosixAttributes` from `WindowsFileSystem`.

This filesystem was borrowed from `CoreRoutingFileSystem` just as a copy-paste. Although the described behavior looks like a bug that can be fixed, the benefits from using `CoreRoutingFileSystem` are not clear either. It's easier to not use it.

GitOrigin-RevId: f734d36907162c83e94db7dd7a569888503a1dda