[terminal] IJPL-159649 Use builtin completion for commands in PowerShell.
For all the Shells we have a custom completion engine, which, as a part of work, collects all the available commands in the Shell. On Windows, the list of commands (applications. cmdlets, functions) could be enormous.
This could sometimes take dozens of seconds to load. This is unacceptable for the completion.
Previously, we applied a workaround to load only imported commands, but this drops many suggestions in the completion.
As a second attempt to make fast and reliable completion for PowerShell, we now delegate the completion to itself, assuming that it provides all possible options to the user.

Now we return emptyList here, to force the fallback to the PowerShell completion, which is executed only if our completion returns no suggestions.


Merge-request: IJ-MR-142490
Merged-by: Vladimir Shefer <Vladimir.Shefer@jetbrains.com>

GitOrigin-RevId: e5ed0e31b024fefb49c7134cf52eabf5f854ec71