[rdct-tests] use custom focus request logic for windows
On windows project window has `isFocused` = true

when on the recording there could be a start window or some app update window is opened. That's why we are trying as much as we can to actually request focus and it helps with update window, but doesn't help with the start menu yet.

GitOrigin-RevId: c0ab3b8a86344a0540052092e05ed5aa74c47e65