RDCT-1444 Add TaskStorage implementations for frontend.split and backend
`TaskStorage` service is responsible for storing/updating/removing tasks from Rhizome DB.
There are two implementations of the service - backend (available in mononlith and backend.split) and frontend.split

The backend implementation stores `TaskInfoEntity` into a shared partition of DB, so the tasks can be available both on backend and all connected frontends.

Frontend split implementation stores `TaskInfoEntity` into a local partition of DB, so the tasks produced by one of the frontends won't be available for backend or other frontends.
Consider "Code With Me" case here: when one of the guests starts some tasks, it shouldn't be visible to other guests.
However, if the host starts a task - all guests should see it.

GitOrigin-RevId: a020d871d2f5f88809dfed551f511b116e8ff550