RDCT-1444 Report progresses to Rhizome DB in PlatformTaskSupport.kt
If flag `rhizome.progress` is enabled, progress info is going to be reported via Rhizome DB rather than with `progressStarted`.

Under the flag no indicators are created right now - additional support has to be added on the side of `StatusBarImpl`

`PlatformTaskSupport` also subscribes to `taskInfo.statuses` to cancel a task if its status has been set to `CANCELED`. `RUNNING` and `PAUSED` statuses are going to be supported later

GitOrigin-RevId: a7b702eaf20260a641e928a95b413e2ba4c61d3b