[rdct] don't restart the backend during config importing (GTW-7436)
After we started migrating configs for the backend process, it may happen that vmoptions are changed during migration, and in that case the process is restarted. Code which works with the backend process may not be ready for that yet (see GTW-9531), so it's more reliable to disable such restarts for now.

Also, logging is improved to mention whether the restarting really happens or not.

GitOrigin-RevId: 109675c52f8861f1b1a801baa3ca0925fd26fff9