Ijent: Get rid of `CoroutineScope` in `IjentFileSystemApi`: no need to export it to the caller.
Coroutine scope is an implementation detail. It was only used to check if filesystem os open, but as a delegate closing process to the `IjentNioFileSystemProvider`, we can check it to see if file system is open.

Moreover, it obeys the contract of `open` function: it now returns false only if `closed` was called.


Merge-request: IJ-MR-142795
Merged-by: Ilya Kazakevich <ilya.kazakevich@jetbrains.com>

GitOrigin-RevId: 80b6f98204d9de6755a2925545435b17a31578a6