[github] Fix the unordered list of items causing incorrect replacement (IJPL-160209)
The list of reviewers was simply emitted, but this
caused a dumb list updating algorithm to replace some entries in the list
without checking equality.

Reason is that it wants to put selected items at the front of the list.
Now, we emit the full list with those selected items at the front,
which prevents erroneous behavior.

GitOrigin-RevId: da238eec1eebd9428cabd3692a7349ea93324047