IJPL-160617 Only expand the first Project View node if no state was restored
When the Project View is shown for the first time, the first node
is selected and expanded. However, if there was a stored state
before, it's also applied, and then the first node is expanded,
which looks like a bug because the state is different from what was saved.

This only affects the case when there are multiple top-level nodes
in the Project View, and the first one was collapsed in the saved state,
but not everything was collapsed. This isn't applicable to a typical
single-module project usually, as it's uncommon for the user to leave,
say, external libraries expanded but the project module collapsed.
But if there are multiple modules, or if the Packages View was open,
for example, then it's quite possible that the first node was
collapsed.

Fix by only doing this select-and-expand thing when the restored
state was empty. In this case it's typically desirable to have at least
something expanded. But if it wasn't empty, then it's likely
to be exactly what the user wants, so we just leave that state alone.

GitOrigin-RevId: f200e22c4ee279abe5917017e6b9376e583f2ee7