PY-76629: Introduce new argument typeCharIfOnlyOneCompletion for completeBasicAllCarets.
There are a few tests that provides `charToTypeAfterCompletion = "\t"` to complete the first variant.
It makes tests "ditry" - somewhere in tests you have to check that "\t" is present after completion,
somewhere "\t" is not present because it completed the first variant.
This argument allows to avoid adding unnecessary `charToTypeAfterCompletion`
when there is only one completion variant that is already completed.

GitOrigin-RevId: 3ed977930b92cedd04ea73764613a55616c291d4