IJPL-159256 run navbar vm on `Dispatchers.Default`
The bug is reproduced when `NavBarItemComponent.focusItem` does not happen before the update request is received.
Once the navbar has the focus, updates from the data context are ignored.

134c533a93c3ca541f0e3e5d19964ad9343e5a93 changed the dispatcher of the NavBarVmImpl to `Dispatchers.EDT`.
The led to `NavBarService.contextItems` flow being collected on EDT (and replacing the whole model) before
`NavBarItemComponent.focusItem` had a chance to be executed.

GitOrigin-RevId: 2607e69434594155769b84dc22b5231135662fad