[RDCT] GTW-9421: Add extension points for project and application settings
To support synchronization of yet not loaded settings, we need a way to list them. Currently, most of the settings are lazy services, and there is no easy way to list all registered settings. This commit adds two extension points: for application- and project-level settings. This will add an ability to iterate though all settings and read their values from a disk.

GitOrigin-RevId: ef3056e223a0361bfa0dc2498188b401ba983e73