IJPL-160930 fix listener leak in JBTabsImpl
Before this fix, a listener was created as a lambda function and then passed to add `addAWTEventListener`.
As `addAWTEventListener` expects an instance of AWTEventListener, the lambda was wrapped with an adapter by Kotlin compiler.
Thus, the actually passed instance was different from the instance in `listener` variable.
And, consequently, unregistering was not successful because yet another wrapped was passed.


(cherry picked from commit e8b07d7e05542274943614197bbb57c2a998f732)

IJ-CR-143269

GitOrigin-RevId: 0890df641a0f1072f0cf5d96aaad11ecc48df574