IJPL-43686 Work arount macOS mouse cursor issues
Sometimes macOS just ignores the cursor we set.
Everything is called correctly, all the way down to
setCursorOnAppKitThread in CCursorManager.
But it's simply ignored. It happens often when
using rounded corners, as the mouse approaches through
the corner.

Temporarily work around this by calling the native
method over and over again. It doesn't fail often,
so this approach works, except that it's a very ugly
reflection hack. But we need at least some fix
until we have the right JBR API to fix it.
Or until a native fix is found.

GitOrigin-RevId: ae713777d3a3dc9106842b39c626ac80abff9bfc