[plugin model] support required content modules (IJPL-148565)
Now it's possible to mark a content module in plugin.xml as required by adding a 'loading="required"' attribute.

If a required module cannot be loaded, the whole plugin is disabled. To achieve this, edges from a required module to the main plugin module and from the main plugin module to each module that depends on a required module are added in the ModuleGraph. This ensures that when a main plugin module is processed in PluginSetBuilder.computeEnabledModuleMap, all required content modules are already processed, but modules from plugins depending on them aren't processed yet, so all required modules are enabled or disabled along with the main plugin module. An alternative solution would be forming a cycle between all required modules in the graph, but this would require special handling since currently all cycles in the module graph are reported as errors. When we'll be able to use dependencies from iml while loading modules (see IJPL-128), we can use actual dependencies between required content modules instead.

All required content modules use the same classloader as the main module, and their dependencies' classloaders are added as parents to that classloader. This is needed to avoid breaking compatibility with other plugins when we extract a required module from an existing plugin and to avoid IllegalAccessError when accessing package private members between such modules (we have many of such cases, see IJPL-149126).

'on-demand' value of 'loading' attribute is supported in code which reads and writes module descriptors, but during loading, such value is treated the same way as 'optional' for now.

GitOrigin-RevId: e709d59b5a3b09b889ff2957e378b6cb05cea2ac